Add arm64 too as compile target

+++ deps/gframe/generic_duel.cpp
@@ -554,7 +554,7 @@ void GenericDuel::TPResult(DuelPlayer* dp, unsigned char tp) {
 	if(host_info.extra_rules & SPEED_DUEL)
 		opt |= DUEL_SPEED;
 	OCG_Player team = { (int)host_info.start_lp, host_info.start_hand, host_info.draw_count };
-	pduel = mainGame->SetupDuel({ rnd(), opt, team, team });
+	pduel = mainGame->SetupDuel({ (uint32_t)rnd(), opt, team, team });
 	if(!host_info.no_shuffle_deck) {
 		ITERATE_PLAYERS(std::shuffle(dueler.pdeck.main.begin(), dueler.pdeck.main.end(), rnd);)
 	}
